requirements.txt
Django==4.1
pycryptodome==3.15.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
requests==2.28.1
django-filter==22.1
psycopg2==2.9.4
coverage==6.5.0
jsonnet==0.18.0
django-nose==1.4.6
django-rest-swagger==2.2.0
selenium==4.7.2
pre-commit==3.5.0
gitlint
pynose==1.4.8
openpyxl
django-parler==2.3
dj-database-url == 2.1.0
whitenoise==6.5.0
gunicorn==21.2.0
